requirements.txt
./local_wheels/visionai-0.0.6-py3-none-any.whl
google-cloud-resource-manager
termcolor